open_ai: Address review feedback
